Reorganize fixtures in tests
This commit extract the parameter combination dictionary as a fixture, and moves
the creation of the CampaignManager to the conftest.py file to allow access to
it from all test files.
